requirements.txt
asgiref==3.7.2
Django==4.2.2
django-autoslug==1.9.9
django-bootstrap-datepicker-plus==5.0.3
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pydantic==1.10.9
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
